social-intelligence-hub/scraper/requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.32.3
supabase==2.5.3
python-dotenv==1.0.1
azure-ai-textanalytics==5.3.0
feedparser==6.0.11
schedule==1.2.2
praw==7.7.1
